# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: kartochki.online API
  version: 0.1.0
  description: |
    Публичный API сервиса kartochki.online.

    Эта спецификация является источником правды для автогенерации frontend-клиента.
    Реализация backend и этот файл должны оставаться синхронными.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/generations:
    post:
      summary: Запустить генерацию карточек
      description: Создаёт проект, записывает параметры запуска и ставит фоновую задачу генерации в очередь.
      operationId: createGeneration
      tags:
        - Generation
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGenerationRequest'
      responses:
        '202':
          description: Генерация поставлена в очередь
          headers:
            X-Request-ID:
              description: Request identifier for correlating logs.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateGenerationResponse'
        '400':
          description: Ошибка валидации параметров генерации
          headers:
            X-Request-ID:
              description: Request identifier for correlating logs.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Токен отсутствует, истёк или недействителен
          headers:
            X-Request-ID:
              description: Request identifier for correlating logs.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Исходное изображение не найдено
          headers:
            X-Request-ID:
              description: Request identifier for correlating logs.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Текущий billing-лимит не позволяет запустить ещё одну генерацию
          headers:
            X-Request-ID:
              description: Request identifier for correlating logs.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка при создании generation job
          headers:
            X-Request-ID:
              description: Request identifier for correlating logs.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: 'Передавайте токен в заголовке Authorization: Bearer <token>.'
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: not_found
        message:
          type: string
          example: ресурс не найден
        request_id:
          type: string
          example: 01HZY3GQK0M5YAWQ2N6R8T9P1S
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      required:
        - message
      properties:
        field:
          type: string
          example: email
        message:
          type: string
          example: поле обязательно
...
    CreateGenerationRequest:
      type: object
      required:
        - marketplace_id
        - style_id
        - card_type_ids
        - card_count
        - source_asset_id
      properties:
        project_name:
          type: string
          maxLength: 200
          example: Карточки для летней коллекции
        marketplace_id:
          type: string
          example: wildberries
        style_id:
          type: string
          example: clean_catalog
        card_type_ids:
          type: array
          minItems: 1
          items:
            type: string
          example:
            - cover
            - benefits
            - details
        card_count:
          type: integer
          example: 5
        source_asset_id:
          type: string
          format: uuid
    CreateGenerationResponse:
      type: object
      required:
        - generation_id
        - status
      properties:
        generation_id:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - queued
          example: queued
...